+++ AutomatedTesting/Gem/PythonTests/scripting/CMakeLists.txt
@@ -28,5 +28,7 @@ if(PAL_TRAIT_BUILD_TESTS_SUPPORTED AND PAL_TRAIT_BUILD_HOST_TOOLS)
             Legacy::Editor
             AZ::AssetProcessor
             AutomatedTesting.Assets
+        COMPONENT
+            ScriptCanvas
     )
 endif()
