+++ AutomatedTesting/Gem/PythonTests/scripting/TestSuite_Periodic.py
@@ -37,24 +37,6 @@ class TestAutomation(TestAutomationBase):
         from . import Pane_HappyPath_ResizesProperly as test_module
         self._run_test(request, workspace, editor, test_module)
 
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptCanvas_TwoComponents_InteractSuccessfully(self, request, workspace, editor, launcher_platform, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptCanvas_TwoComponents_InteractSuccessfully as test_module
-        self._run_test(request, workspace, editor, test_module)
-
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptCanvas_ChangingAssets_ComponentStable(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptCanvas_ChangingAssets_ComponentStable as test_module
-        self._run_test(request, workspace, editor, test_module)
-
     def test_Graph_HappyPath_ZoomInZoomOut(self, request, workspace, editor, launcher_platform):
         from . import Graph_HappyPath_ZoomInZoomOut as test_module
         self._run_test(request, workspace, editor, test_module)
@@ -63,26 +45,8 @@ class TestAutomation(TestAutomationBase):
         from . import NodePalette_HappyPath_CanSelectNode as test_module
         self._run_test(request, workspace, editor, test_module)
 
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptCanvasComponent_OnEntityActivatedDeactivated_PrintMessage(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptCanvasComponent_OnEntityActivatedDeactivated_PrintMessage as test_module
-        self._run_test(request, workspace, editor, test_module)
-
     def test_NodePalette_HappyPath_ClearSelection(self, request, workspace, editor, launcher_platform, project):
         from . import NodePalette_HappyPath_ClearSelection as test_module
-        self._run_test(request, workspace, editor, test_module)
-
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptCanvas_TwoEntities_UseSimultaneously(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptCanvas_TwoEntities_UseSimultaneously as test_module
         self._run_test(request, workspace, editor, test_module)
 
     def test_ScriptEvent_HappyPath_CreatedWithoutError(self, request, workspace, editor, launcher_platform, project):
@@ -139,33 +103,6 @@ class TestAutomation(TestAutomationBase):
         from . import Pane_Default_RetainOnSCRestart as test_module
         self._run_test(request, workspace, editor, test_module)
 
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptEvents_HappyPath_SendReceiveAcrossMultiple(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptEvents_HappyPath_SendReceiveAcrossMultiple as test_module
-        self._run_test(request, workspace, editor, test_module)
-
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptEvents_Default_SendReceiveSuccessfully(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptEvents_Default_SendReceiveSuccessfully as test_module
-        self._run_test(request, workspace, editor, test_module)
-
-    @pytest.mark.parametrize("level", ["tmp_level"])
-    def test_ScriptEvents_ReturnSetType_Successfully(self, request, workspace, editor, launcher_platform, project, level):
-        def teardown():
-            file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        request.addfinalizer(teardown)
-        file_system.delete([os.path.join(workspace.paths.project(), "Levels", level)], True, True)
-        from . import ScriptEvents_ReturnSetType_Successfully as test_module
-        self._run_test(request, workspace, editor, test_module)
-
     def test_NodeCategory_ExpandOnClick(self, request, workspace, editor, launcher_platform):
         from . import NodeCategory_ExpandOnClick as test_module
         self._run_test(request, workspace, editor, test_module)
@@ -203,15 +140,6 @@ class TestScriptCanvasTests(object):
     """
     The following tests use hydra_test_utils.py to launch the editor and validate the results.
     """
-
-    def test_FileMenu_Default_NewAndOpen(self, request, editor, launcher_platform):
-        expected_lines = [
-            "File->New action working as expected: True",
-            "File->Open action working as expected: True",
-        ]
-        hydra.launch_and_validate_results(
-            request, TEST_DIRECTORY, editor, "FileMenu_Default_NewAndOpen.py", expected_lines, auto_test_mode=False, timeout=60,
-        )
 
     def test_NewScriptEventButton_HappyPath_ContainsSCCategory(self, request, editor, launcher_platform):
         expected_lines = [
@@ -282,7 +210,6 @@ class TestScriptCanvasTests(object):
             },
         ],
     )
-
     def test_Pane_PropertiesChanged_RetainsOnRestart(self, request, editor, config, project, launcher_platform):
         hydra.launch_and_validate_results(
             request,
@@ -291,59 +218,6 @@ class TestScriptCanvasTests(object):
             "Pane_PropertiesChanged_RetainsOnRestart.py",
             config.get('expected_lines'),
             cfg_args=[config.get('cfg_args')],
-            auto_test_mode=False,
-            timeout=60,
-        )
-
-    def test_ScriptEvent_AddRemoveMethod_UpdatesInSC(self, request, workspace, editor, launcher_platform):
-        def teardown():
-            file_system.delete(
-                [os.path.join(workspace.paths.project(), "TestAssets", "test_file.scriptevents")], True, True
-            )
-        request.addfinalizer(teardown)
-        file_system.delete(
-            [os.path.join(workspace.paths.project(), "TestAssets", "test_file.scriptevents")], True, True
-        )
-        expected_lines = [
-            "Success: New Script Event created",
-            "Success: Initial Child Event created",
-            "Success: Second Child Event created",
-            "Success: Script event file saved",
-            "Success: Method added to scriptevent file",
-            "Success: Method removed from scriptevent file",
-        ]
-        hydra.launch_and_validate_results(
-            request,
-            TEST_DIRECTORY,
-            editor,
-            "ScriptEvent_AddRemoveMethod_UpdatesInSC.py",
-            expected_lines,
-            auto_test_mode=False,
-            timeout=60,
-        )
-
-    def test_ScriptEvents_AllParamDatatypes_CreationSuccess(self, request, workspace, editor, launcher_platform):
-        def teardown():
-            file_system.delete(
-                [os.path.join(workspace.paths.project(), "TestAssets", "test_file.scriptevents")], True, True
-            )
-        request.addfinalizer(teardown)
-        file_system.delete(
-            [os.path.join(workspace.paths.project(), "TestAssets", "test_file.scriptevents")], True, True
-        )
-        expected_lines = [
-            "Success: New Script Event created",
-            "Success: Child Event created",
-            "Success: New parameters added",
-            "Success: Script event file saved",
-            "Success: Node found in Script Canvas",
-        ]
-        hydra.launch_and_validate_results(
-            request,
-            TEST_DIRECTORY,
-            editor,
-            "ScriptEvents_AllParamDatatypes_CreationSuccess.py",
-            expected_lines,
             auto_test_mode=False,
             timeout=60,
         )
